input: Refactor ARM9 input handling
Removed input fields from transfer region. Instead, the FIFO handler saves
the last received values, and scanKeys() refreshes the set of values that
touchPosition() returns.

This prevents a race condition when keysHeld() may say that KEY_TOUCH is
pressed, but the touch screen status is updated before calling
touchPosition(), which returns (0, 0) because the player has stopped
touching the screen.

Also, interrupts need to be off while accessing the latched data because
it may change at any point without warning from an interrupt handler.
